moved prediction to separate file

+++ src/epos/clock.hpp
@@ -6,13 +6,13 @@ namespace epos {
 using namespace std::chrono_literals;
 
 using clock = std::chrono::high_resolution_clock;
-using nanoseconds = std::chrono::duration<double, std::chrono::nanoseconds::period>;
-using microseconds = std::chrono::duration<double, std::chrono::microseconds::period>;
-using milliseconds = std::chrono::duration<double, std::chrono::milliseconds::period>;
-using seconds = std::chrono::duration<double, std::chrono::seconds::period>;
-using minutes = std::chrono::duration<double, std::chrono::minutes::period>;
-using hours = std::chrono::duration<double, std::chrono::hours::period>;
-using days = std::chrono::duration<double, std::chrono::days::period>;
+using nanoseconds = std::chrono::duration<float, std::chrono::nanoseconds::period>;
+using microseconds = std::chrono::duration<float, std::chrono::microseconds::period>;
+using milliseconds = std::chrono::duration<float, std::chrono::milliseconds::period>;
+using seconds = std::chrono::duration<float, std::chrono::seconds::period>;
+using minutes = std::chrono::duration<float, std::chrono::minutes::period>;
+using hours = std::chrono::duration<float, std::chrono::hours::period>;
+using days = std::chrono::duration<float, std::chrono::days::period>;
 
 using std::chrono::duration_cast;
 
@@ -24,22 +24,22 @@ struct interval {
     return clock::now() - start;
   }
 
-  double s() const noexcept
+  float s() const noexcept
   {
     return duration_cast<seconds>(get()).count();
   }
 
-  double ms() const noexcept
+  float ms() const noexcept
   {
     return duration_cast<milliseconds>(get()).count();
   }
 
-  double us() const noexcept
+  float us() const noexcept
   {
     return duration_cast<microseconds>(get()).count();
   }
 
-  double ns() const noexcept
+  float ns() const noexcept
   {
     return duration_cast<nanoseconds>(get()).count();
   }
